refactor third_party
